Add option for align repos's description
The option is `helm-github-stars-name-length', its default value
is 30.

There is a tiny change that user opens a wrong URL when both
description and (substring name 0 helm-github-stars-name-length)
is not unique. e.g.,

username/repo-foo - This is a sample repo
username/repo-bar - This is a sample repo

when helm-github-stars-name-length is 10, Helm show the following:

username/repo...   This is a sample repo
username/repo...   This is a sample repo

in this case the result might get wrong.

Fixes #8